+++ src/MaterialsLibrary/src/materialslibrary/simple/simple_material.cpp
@@ -215,9 +215,9 @@ void SimpleMaterial::bindForSubMesh(Matrix* world, Mesh* mesh, SubMesh* subMesh)
     if (_diffuseTexture && StandardMaterial::DiffuseTextureEnabled()) {
       _activeEffect->setTexture("diffuseSampler", _diffuseTexture);
 
-      _activeEffect->setFloat2("vDiffuseInfos",
-                               _diffuseTexture->coordinatesIndex,
-                               _diffuseTexture->level);
+      _activeEffect->setFloat2(
+        "vDiffuseInfos", static_cast<float>(_diffuseTexture->coordinatesIndex),
+        _diffuseTexture->level);
       _activeEffect->setMatrix("diffuseMatrix",
                                *_diffuseTexture->getTextureMatrix());
     }
