+++ src/ProceduralTexturesLibrary/src/proceduraltextureslibrary/perlinnoise/perlin_noise_procedural_texture.cpp
@@ -29,7 +29,7 @@ PerlinNoiseProceduralTexture::~PerlinNoiseProceduralTexture()
 
 void PerlinNoiseProceduralTexture::updateShaderUniforms()
 {
-  setFloat("size", getRenderSize().width);
+  setFloat("size", static_cast<float>(getRenderSize().width));
 
   auto deltaTime = getScene()->getEngine()->getDeltaTime();
